backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Magnesium- <yyong007@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
zhipuai = "^2.1.4.20230814"
supabase = "^2.7.3"
langchain = "^0.3.3"
langchainhub = "^0.1.21"
langchain-openai = "^0.2.2"
langchain-community = "^0.3.3"
sentence-transformers = "^3.2.1"
faiss-cpu = "^1.9.0"
langchain-chroma = "^0.1.4"
langgraph = "^0.2.39"
bs4 = "^0.0.2"
python-multipart = "^0.0.17"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
